Feature(Cancel booking): update

+++ Core/Fieldy.BookingYard.Domain/Entities/HistoryPoint.cs
@@ -8,7 +8,7 @@ namespace Fieldy.BookingYard.Domain.Entities
 	{
         public Guid UserID { get; set; }
         public User? User{ get; set; }
-        public decimal Point { get; set; }
+        public int Point { get; set; }
         public DateTime CreatedAt { get; set; }
         public string? Content { get; set; }
     }
